refactor: reorganize charge sensor test fixtures and enhance modularity
- Moved charge sensor-related test fixtures from `tests/conftest.py` to a new dedicated `tests/routines/builtins/charge_sensor/conftest.py` file for better organization.
- Removed unused imports and fixtures from the original test configuration to streamline the codebase.
- Introduced `model_fit_result`, `fitted_peak_factory`, and `stability_measurement_factory` fixtures in the new file to facilitate testing of charge sensor routines.
- Improved overall structure of test files for enhanced maintainability and clarity.
